Restructure 2d view
