+++ src/main/java/meshIneBits/gui/utilities/IconLoader.java
@@ -20,7 +20,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package meshIneBits.gui;
+package meshIneBits.gui.utilities;
 
 import javax.swing.*;
 import java.awt.*;
