+++ src/main/java/meshIneBits/gui/utilities/ToggleIcon.java
@@ -22,8 +22,8 @@
 
 package meshIneBits.gui.utilities;
 
-import meshIneBits.gui.MeshAction;
-import meshIneBits.gui.MeshToggleAction;
+import meshIneBits.gui.view2d.MeshAction;
+import meshIneBits.gui.view2d.MeshToggleAction;
 
 import javax.swing.*;
 import javax.swing.border.Border;
