+++ src/main/java/meshIneBits/gui/utilities/UPPPaveRegion.java
@@ -23,8 +23,8 @@
 package meshIneBits.gui.utilities;
 
 import meshIneBits.config.CraftConfig;
-import meshIneBits.gui.MeshController;
-import meshIneBits.gui.MeshToggleAction;
+import meshIneBits.gui.view2d.MeshController;
+import meshIneBits.gui.view2d.MeshToggleAction;
 
 import javax.swing.*;
 import java.awt.*;
