+++ src/main/java/meshIneBits/gui/view2d/MeshAction.java
@@ -20,7 +20,9 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package meshIneBits.gui;
+package meshIneBits.gui.view2d;
+
+import meshIneBits.gui.utilities.IconLoader;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
