+++ src/main/java/meshIneBits/gui/view2d/MeshToggleAction.java
@@ -20,7 +20,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package meshIneBits.gui;
+package meshIneBits.gui.view2d;
 
 public class MeshToggleAction extends MeshAction {
