+++ src/main/java/meshIneBits/gui/view2d/MeshWindow.java
@@ -20,7 +20,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package meshIneBits.gui;
+package meshIneBits.gui.view2d;
 
 import meshIneBits.config.CraftConfig;
 import meshIneBits.config.CraftConfigLoader;
